requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
bson==0.5.10
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.7.0
executing==2.2.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.4.0
matplotlib-inline==0.1.7
maturin==1.8.3
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.5
odfpy==1.4.1
packaging==25.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pymongo==4.12.0
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
threadpoolctl==3.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
underthesea==6.8.4
underthesea_core==1.0.4
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
